+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.2.4
 * [ADD] Implementation of automated tests for drawings
+* [FIX] Fix the fontweight of "big basic" font
 
 ## 0.2.3
 * [IMP] Improve the Bluetooth status detection